+++ setup.cfg
@@ -65,7 +65,7 @@ markers =
     unit: marks tests as unit (deselect with '-m "not unit"')
 # -- recommended but optional:
 python_files = tests.py test_*.py *_tests.py
-addopts = -vv -s --disable-warnings --reuse-db
+addopts = -s -x --disable-warnings --reuse-db
 filterwarnings =
     ignore::DeprecationWarning
     ignore::PendingDeprecationWarning
